Raf aligned dom updates (#604)
* Raf aligned dom updates

* Test approach leveraging our existing options.debounceRendering

This all works fine, however it creates a discrepancy between
browser usage and test usage which we want to avoid. The discrepancy
is made clear in the failing tests.

What we want to happen is for Preact to be first allowed to perform
its render cycle, the effects that follow and only then should we
clean up stragglers by performing the direct DOM updates. Currently
what happens is that the ordering is "random" based on how the signals
were registered.

* options.requestAnimationFrame solves it

* Improve changelog post

Co-authored-by: Ryan Christian <33403762+rschristian@users.noreply.github.com>

* Update .changeset/ninety-beans-compare.md

* Skip changing the SignalValue notifier when it's an object (JSX)

---------

Co-authored-by: Ryan Christian <33403762+rschristian@users.noreply.github.com>